refactor: Move fragment to Status package
