+++ GPSTest/src/main/java/com/android/gpstest/ui/MainActivity.kt
@@ -57,6 +57,7 @@ import com.android.gpstest.data.LocationRepository
 import com.android.gpstest.databinding.ActivityMainBinding
 import com.android.gpstest.map.MapConstants
 import com.android.gpstest.ui.NavigationDrawerFragment.NavigationDrawerCallbacks
+import com.android.gpstest.ui.status.StatusFragment2
 import com.android.gpstest.util.*
 import com.android.gpstest.util.PreferenceUtils.isTrackingStarted
 import com.android.gpstest.util.SharedPreferenceUtil.darkTheme
